requirements.txt
imgaug==0.4.0
numpy==1.21.5
Pillow==9.5.0
